+++ dcgpy/expose_kernels.cpp
@@ -9,75 +9,19 @@
 #include <vector>
 
 #include <dcgp/function.hpp>
-#include <dcgp/wrapped_functions_s11n_implement.hpp>
 #include <dcgp/kernel.hpp>
 #include <dcgp/kernel_set.hpp>
 #include <dcgp/s11n.hpp>
+#include <dcgp/wrapped_functions_s11n_implement.hpp>
 
 #include "common_utils.hpp"
 #include "docstrings.hpp"
+#include "pybind11_function.hpp"
 
 using namespace dcgp;
 using namespace dcgpy;
 using namespace audi;
 namespace py = pybind11;
-
-namespace dcgp::detail
-{
-
-template <typename T>
-struct function_inner<py::object, T, const std::vector<T> &> final : function_inner_base<T, const std::vector<T> &> {
-    // We just need the def ctor, delete everything else.
-    function_inner() = default;
-    function_inner(const function_inner &) = delete;
-    function_inner(function_inner &&) = delete;
-    function_inner &operator=(const function_inner &) = delete;
-    function_inner &operator=(function_inner &&) = delete;
-
-    // Constructor from generic python object.
-    explicit function_inner(const py::object &o)
-    {
-        m_value = dcgpy::deepcopy(o);
-    }
-
-    // Clone method.
-    virtual std::unique_ptr<function_inner_base<T, const std::vector<T> &>> clone() const override final
-    {
-        // This will make a deep copy using the ctor above.
-        return std::make_unique<function_inner>(m_value);
-    }
-
-    // Mandatory methods.
-    virtual T operator()(const std::vector<T> &v) const override final
-    {
-        return py::cast<T>(m_value(v));
-    }
-
-    virtual pagmo::thread_safety get_thread_safety() const override final
-    {
-        return pagmo::thread_safety::none;
-    }
-
-    template <typename Archive>
-    void save(Archive &ar, unsigned) const
-    {
-        ar << boost::serialization::base_object<function_inner_base<T, const std::vector<T> &>>(*this);
-        ar << dcgpy::object_to_vchar(m_value);
-    }
-    template <typename Archive>
-    void load(Archive &ar, unsigned)
-    {
-        ar >> boost::serialization::base_object<function_inner_base<T, const std::vector<T> &>>(*this);
-        std::vector<char> v;
-        ar >> v;
-        m_value = dcgpy::vchar_to_object(v);
-    }
-    BOOST_SERIALIZATION_SPLIT_MEMBER()
-
-    py::object m_value;
-};
-
-} // namespace dcgp::detail
 
 namespace dcgp::s11n_names
 {
@@ -228,13 +172,12 @@ void expose_kernel_set(const py::module &m, std::string type)
     ker_set.def(py::init<>())
         .def(py::init<std::vector<std::string>>(), py::arg("kernels"), kernel_set_init_doc(type).c_str())
         .def("__call__", &kernel_set<T>::operator())
-        .def(
-            "__repr__",
-            [](const kernel_set<T> &ks) -> std::string {
-                std::ostringstream oss;
-                oss << ks;
-                return oss.str();
-            })
+        .def("__repr__",
+             [](const kernel_set<T> &ks) -> std::string {
+                 std::ostringstream oss;
+                 oss << ks;
+                 return oss.str();
+             })
         .def(
             "push_back", [](kernel_set<T> &ks, std::string v) { ks.push_back(v); },
             kernel_set_push_back_str_doc().c_str(), py::arg("kernel_name"))
